+++ Biit Employee Performance Apraisal API/SubkpiEmployeeScore.cs
@@ -18,7 +18,7 @@ namespace Biit_Employee_Performance_Apraisal_API
         public int subkpi_id { get; set; }
         public int employee_id { get; set; }
         public int session_id { get; set; }
-        public Nullable<double> score { get; set; }
+        public Nullable<int> score { get; set; }
 
         [JsonIgnore]
         public virtual Employee Employee { get; set; }
